build(vcpkg): migrate bundled third-party deps to vcpkg
Replace the bundled Eigen / Ceres 1.14 / GLFW / spdlog(1.5, fmt 5.5) /
pugixml under src/Third-party and src/LibGnut with vcpkg-managed packages
declared in src/vcpkg.json (ceres, eigen3, glfw3, pugixml, spdlog). This
removes the vendored fmt 5.5 that fails MSVC 14.51 (VS2026) with
C2143/C2447.

- Pin eigen3 to 3.4.1 (vcpkg overrides) and Ceres to 2.2.0; build C++17
  (Ceres 2.x requirement) with _HAS_STD_BYTE=0 to avoid the Windows SDK
  std::byte / objidl.h collision under C++17.
- Point CMake at vcpkg imported targets; switch pugixml includes from
  "pugixml/src/pugixml.hpp" to "pugixml.hpp"; cast the input-format enum to
  int in spdlog::critical for modern fmt.
- Add CMakePresets.json (VS2026 + vcpkg), src/vcpkg-configuration.json, and
  an eigen3 overlay port (tools/ports/eigen3) pinning a refreshed SHA512 for
  GitLab's non-reproducible eigen archive tarball.
- Stage vcpkg runtime DLLs beside each app EXE via TARGET_RUNTIME_DLLS.
- Ignore build_vcpkg/ and vcpkg_installed/.

+++ src/LibGnut/gset/gsetinp.cpp
@@ -94,7 +94,7 @@ namespace gnut
 		case IFMT::EOP_INP:
 			return "EOP";
         default:
-            spdlog::critical("No fmt for {}, check your inp.", f);
+            spdlog::critical("No fmt for {}, check your inp.", static_cast<int>(f));
             throw logic_error("check your inp");
         }
     }
